backend/requirements.txt
Flask==3.1.0
Werkzeug==3.1.3
SQLAlchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
boto3==1.35.82
confluent-kafka==2.6.1
firebase-admin==6.6.0
google-auth==2.37.0
openpyxl==3.1.5
pdfplumber==0.11.4
PyMuPDF==1.25.1
python-dotenv==1.0.1
pytest==8.3.4
requests==2.32.3
black==24.10.0
ruff==0.7.4
mypy==1.13.0
